requirements.txt
# QuantMCP 项目依赖
python-dotenv>=1.0.0

# FastMCP 2.0库
fastmcp>=2.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 迅投量化
xtquant

# 数学计算
scipy>=1.10.0
scikit-learn>=1.3.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 